+++ tests/testthat/test-CheckDrugSig.R
@@ -13,20 +13,28 @@ getTestData <- function() {
 
 test_that("checkDrugSig overall", {
   testData <- getTestData()
-  result <- checkDrugSig(testData, byConcept = FALSE) %>% dplyr::collect()
+  result <- checkDrugSig(testData, byConcept = FALSE) %>%
+    dplyr::collect() %>%
+    dplyr::mutate(ingredient_concept_id = as.numeric(.data$ingredient_concept_id)) %>%
+    dplyr::arrange(.data$ingredient_concept_id, dplyr::desc(.data$n_records))
 
   expect_equal(nrow(result), 5)
   expect_equal(ncol(result), 4)
-  expect_equal(result$sig, c("sig1", "sig2", "sig3", "sig1", NA))
-  expect_equal(result$n_records, c(1, 2, 1, 2, 1))
+  expect_equal(result$ingredient_concept_id, c(1, 1, 2, 2, 3))
+  expect_equal(result$sig, c("sig2", "sig1", "sig1", "sig3", NA))
+  expect_equal(result$n_records, c(2, 1, 2, 1, 1))
 })
 
 test_that("checkDrugSig byConcept", {
   testData <- getTestData()
-  result <- checkDrugSig(testData, byConcept = TRUE) %>% dplyr::collect()
+  result <- checkDrugSig(testData, byConcept = TRUE) %>%
+    dplyr::collect() %>%
+    dplyr::mutate(ingredient_concept_id = as.numeric(.data$ingredient_concept_id)) %>%
+    dplyr::arrange(.data$ingredient_concept_id, dplyr::desc(.data$n_records))
 
   expect_equal(nrow(result), 7)
   expect_equal(ncol(result), 6)
-  expect_equal(result$sig, c("sig1", "sig2", "sig3", "sig1", NA, "sig1", "sig2"))
+  expect_equal(result$ingredient_concept_id, c(1, 1, 1, 2, 2, 2, 3))
+  expect_equal(setdiff(result$sig, c("sig1", "sig2", "sig2", "sig1", "sig1", "sig3", NA)), character(0))
   expect_equal(result$n_records, c(1, 1, 1, 1, 1, 1, 1))
 })